requirements.txt
accelerate==1.8.1
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
diffusers==0.34.0
fastapi==0.116.1
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
packaging==25.0
Pillow==11.3.0
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scipy==1.16.0
setuptools==80.9.0
sniffio==1.3.1
starlette==0.47.1
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
torchaudio==2.2.2
tqdm==4.67.1
transformers==4.53.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
zipp==3.23.0
audiocraft==1.0.0
imageio==2.34.1
librosa==0.10.2.post1